changed sized types to unsized types
git-svn-id: svn+ssh://svn.code.sf.net/p/gaia-clean/code/gaia/branches/gaia-tiledb@68 6c2f56dd-5223-0410-8aad-addf9322a0be

+++ lib/tiledb/tiledb_internal.h
@@ -7,8 +7,8 @@
 void create_new_db(DB_Handle* db_handle);
 
 void init_index_entry_v0(tiledb_index_entry_v0 *entry);
-uint32_t get_next_offset_v0(tiledb_index_entry_v0 *entry, uint32_t start_level, uint32_t x, uint32_t y, uint32_t level);
-tiledb_error tiledb_put_v0(DB_Handle* db_handle, uint32_t x, uint32_t y, uint32_t level, char* data, uint32_t size);
-tiledb_error tiledb_get_v0(DB_Handle* db_handle, uint32_t x, uint32_t y, uint32_t level);
+uint32_t get_next_offset_v0(tiledb_index_entry_v0 *entry, unsigned int start_level, unsigned int x, unsigned int y, unsigned int level);
+tiledb_error tiledb_put_v0(DB_Handle* db_handle, unsigned int x, unsigned int y, unsigned int level, char* data, size_t size);
+tiledb_error tiledb_get_v0(DB_Handle* db_handle, unsigned int x, unsigned int y, unsigned int level);
 
 #endif
